LayoutTests/accessibility/alt-tag-on-image-with-nonimage-role.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
<head>
<script src="../resources/js-test-pre.js"></script>
</head>
<body id="body">

<img alt="TEST1" src="resources/cake.png" role="text" id="text">

<img alt="TEST2" src="resources/cake.png" role="group" id="group">

<img alt="TEST3" src="resources/cake.png" role="button" id="button">

<p id="description"></p>
<div id="console"></div>

<script>

    description("This tests that setting a role on an img still allows the alt attribute to be used for the description..");

    if (window.accessibilityController) {

        var text = accessibilityController.accessibleElementById("text");
        shouldBe("text.stringValue", "'AXValue: TEST1'");

        var group = accessibilityController.accessibleElementById("group");
        var button = accessibilityController.accessibleElementById("button");

        if (accessibilityController.platformName == "atk") {
            shouldBe("group.title", "'AXTitle: TEST2'");
            shouldBe("button.title", "'AXTitle: TEST3'");
        } else {
            shouldBe("group.description", "'AXDescription: TEST2'");
            shouldBe("button.description", "'AXDescription: TEST3'");
        }
    }

</script>

<script src="../resources/js-test-post.js"></script>
</body>
</html>
